Refactors bin script -> bin & lib scripts
Separated our concerns by keeping the bin/ script specific to everything
that has to do with the command line, and the lib/ script being
everything related to the upload. Now the s3Upload function (that get's)
exported from our lib/ script just requires an object to be passed to
it with `.path` defined, and it will do the rest.

Fun note: added a .gif ;-)